chore: relocate API v1 router
